feat!: add support for Node.js v16 (#28)
* feat!: add support for Node.js v16

ABI doesn't seem to have changed significantly between 14 and 16, so no
changes are necessary on that front. But V8 is using C++14 features now
so we need to change the build parameters to use C++14. Because this
increases the minimum C compiler version needed to build this package,
this is a breaking change.

Ref:
https://github.com/nodejs/node-gyp/issues/2384#issuecomment-826131823

* 1.0.0

* chore: remove lock file

Lockfiles are not recommended in libraries.